#1559 pull out click to actions
