+++ lib/actions/pageActionChecks.js
@@ -1,4 +1,5 @@
 const { isElement, isSelector, isString, waitUntil } = require('../helper');
+const runtimeHandler = require('../handlers/runtimeHandler');
 const { findElements } = require('../elementSearch');
 const { defaultConfig } = require('../config');
 
@@ -8,6 +9,53 @@ const checkVisible = async (elem) => {
 const checkNotDisabled = async (elem) => {
   return !(await elem.isDisabled());
 };
+
+const checkIfElementIsCovered = async (e) => {
+  function isElementAtPointOrChild() {
+    function getDirectParent(nodes, elem) {
+      return nodes.find((node) => node.contains(elem));
+    }
+
+    let value,
+      elem = this;
+    if (elem.nodeType === Node.TEXT_NODE) {
+      let range = document.createRange();
+      range.selectNodeContents(elem);
+      value = range.getClientRects()[0];
+      elem = elem.parentElement;
+    } else {
+      value = elem.getBoundingClientRect();
+    }
+    const y = (value.top + value.bottom) / 2;
+    const x = (value.left + value.right) / 2;
+
+    let nodes = document.elementsFromPoint(x, y);
+    const parsedShadowRoots = [];
+    while (nodes[0].shadowRoot && !parsedShadowRoots.includes(nodes[0].shadowRoot)) {
+      parsedShadowRoots.push(nodes[0].shadowRoot);
+      nodes = [nodes[0].shadowRoot.elementFromPoint(x, y)];
+    }
+    const isElementCoveredByAnotherElement = nodes[0] !== elem;
+    let node = null;
+    if (isElementCoveredByAnotherElement) {
+      node = nodes[0];
+    } else {
+      node = getDirectParent(nodes, elem);
+    }
+    return (
+      elem.contains(node) ||
+      node.contains(elem) ||
+      window.getComputedStyle(node).getPropertyValue('opacity') < 0.1 ||
+      window.getComputedStyle(elem).getPropertyValue('opacity') < 0.1
+    );
+  }
+  const objectId = e.get();
+  const res = await runtimeHandler.runtimeCallFunctionOn(isElementAtPointOrChild, null, {
+    objectId: objectId,
+  });
+  return res.result.value;
+};
+
 const checksMap = {
   visible: checkVisible,
   disabled: checkNotDisabled,
@@ -70,4 +118,4 @@ const description = (selector, lowerCase = false) => {
   return lowerCase ? d.charAt(0).toLowerCase() + d.slice(1) : d;
 };
 
-module.exports = { description, waitAndGetActionableElement };
+module.exports = { description, waitAndGetActionableElement, checkIfElementIsCovered };
