+++ lib/handlers/inputHandler.js
@@ -163,6 +163,15 @@ const tap = async (x, y) => {
   });
 };
 
+const simulateMouseClick = async (options) => {
+  options.type = 'mouseMoved';
+  await input.dispatchMouseEvent(options);
+  options.type = 'mousePressed';
+  await input.dispatchMouseEvent(options);
+  options.type = 'mouseReleased';
+  await input.dispatchMouseEvent(options);
+};
+
 const _insertText = async (char) => {
   await input.insertText({ text: char });
 };
@@ -177,5 +186,6 @@ module.exports = {
   mouse_move: _mouse_move,
   sendCharacter: _insertText,
   tap,
+  simulateMouseClick,
   dispatchMouseEvent,
 };
