+++ lib/taiko.js
@@ -24,7 +24,7 @@ const targetHandler = require('./handlers/targetHandler');
 const runtimeHandler = require('./handlers/runtimeHandler');
 const browserHandler = require('./handlers/browserHandler');
 const emulationHandler = require('./handlers/emulationHandler');
-const { description } = require('./actions/pageActionChecks');
+const { description, checkIfElementIsCovered } = require('./actions/pageActionChecks');
 const { logEvent } = require('./logger');
 const { match, $$xpath, findElements, findFirstElement } = require('./elementSearch');
 const { handleRelativeSearch, RelativeSearchElement } = require('./proximityElementSearch');
@@ -1113,107 +1113,6 @@ const title = async () => {
 };
 module.exports.title = title;
 
-const checkIfElementIsCovered = async (e) => {
-  function isElementAtPointOrChild() {
-    function getDirectParent(nodes, elem) {
-      return nodes.find((node) => node.contains(elem));
-    }
-
-    let value,
-      elem = this;
-    if (elem.nodeType === Node.TEXT_NODE) {
-      let range = document.createRange();
-      range.selectNodeContents(elem);
-      value = range.getClientRects()[0];
-      elem = elem.parentElement;
-    } else {
-      value = elem.getBoundingClientRect();
-    }
-    const y = (value.top + value.bottom) / 2;
-    const x = (value.left + value.right) / 2;
-
-    let nodes = document.elementsFromPoint(x, y);
-    const parsedShadowRoots = [];
-    while (nodes[0].shadowRoot && !parsedShadowRoots.includes(nodes[0].shadowRoot)) {
-      parsedShadowRoots.push(nodes[0].shadowRoot);
-      nodes = [nodes[0].shadowRoot.elementFromPoint(x, y)];
-    }
-    const isElementCoveredByAnotherElement = nodes[0] !== elem;
-    let node = null;
-    if (isElementCoveredByAnotherElement) {
-      node = nodes[0];
-    } else {
-      node = getDirectParent(nodes, elem);
-    }
-    return (
-      elem.contains(node) ||
-      node.contains(elem) ||
-      window.getComputedStyle(node).getPropertyValue('opacity') < 0.1 ||
-      window.getComputedStyle(elem).getPropertyValue('opacity') < 0.1
-    );
-  }
-  const objectId = e.get();
-  const res = await runtimeHandler.runtimeCallFunctionOn(isElementAtPointOrChild, null, {
-    objectId: objectId,
-  });
-  return res.result.value;
-};
-
-async function _click(selector, options, ...args) {
-  const elems = await handleRelativeSearch(await findElements(selector), args);
-  let elemsLength = elems.length;
-  let isElemAtPoint;
-  let X;
-  let Y;
-  options = setClickOptions(options);
-  if (elemsLength > options.elementsToMatch) {
-    elems.splice(options.elementsToMatch, elems.length);
-  }
-  for (let elem of elems) {
-    const objectId = elem.get();
-    isElemAtPoint = false;
-    await scrollToElement(elem);
-    let { x, y } = await domHandler.boundingBoxCenter(objectId);
-    (X = x), (Y = y);
-    isElemAtPoint = await checkIfElementIsCovered(elem);
-    let isDisabled = (
-      await evaluate(elem, function () {
-        return this.hasAttribute('disabled') ? this.disabled : false;
-      })
-    ).value;
-    if (isDisabled) {
-      throw Error(description(selector) + 'is disabled');
-    }
-    if (isElemAtPoint) {
-      const type = (
-        await evaluate(elem, function getType() {
-          return this.type;
-        })
-      ).value;
-      assertType(
-        objectId,
-        () => type !== 'file',
-        'Unsupported operation, use `attach` on file input field',
-      );
-      if (defaultConfig.headful) {
-        await highlightElement(elem);
-      }
-      break;
-    }
-  }
-  if (!isElemAtPoint && elemsLength != elems.length) {
-    throw Error('Please provide a more specific selector, too many matches.');
-  }
-  if (!isElemAtPoint) {
-    throw Error(description(selector) + ' is covered by other element');
-  }
-  (options.x = X), (options.y = Y);
-  options.noOfClicks = options.noOfClicks || 1;
-  for (let count = 0; count < options.noOfClicks; count++) {
-    await waitForMouseActions(options);
-  }
-}
-
 /**
  * Fetches an element with the given selector, scrolls it into view if needed, and then clicks in the center of the element. If there's no element matching selector, the method throws an error.
  * @example
@@ -1240,38 +1139,12 @@ async function _click(selector, options, ...args) {
  *
  * @returns {Promise<void>}
  */
-module.exports.click = click;
-
-async function click(selector, options = {}, ...args) {
+module.exports.click = async (selector, options = {}, ...args) => {
   validate();
-  if (options instanceof RelativeSearchElement) {
-    args = [options].concat(args);
-    options = {};
-  }
-  if (isSelector(selector) || isString(selector) || isElement(selector)) {
-    options.noOfClicks = options.clickCount || 1;
-    await _click(selector, options, ...args);
-    descEvent.emit(
-      'success',
-      'Clicked ' + description(selector, true) + ' ' + options.noOfClicks + ' times',
-    );
-  } else {
-    options = setClickOptions(options, selector.x, selector.y);
-    options.noOfClicks = options.clickCount || 1;
-    for (let count = 0; count < options.noOfClicks; count++) {
-      await waitForMouseActions(options);
-    }
-    descEvent.emit(
-      'success',
-      'Clicked ' +
-        options.noOfClicks +
-        ' times on coordinates x : ' +
-        selector.x +
-        ' and y : ' +
-        selector.y,
-    );
-  }
-}
+  const { click } = require('./actions/click');
+  const desc = await click(selector, options, ...args);
+  descEvent.emit('success', desc);
+};
 
 /**
  * Fetches an element with the given selector, scrolls it into view if needed, and then double clicks the element. If there's no element matching selector, the method throws an error.
@@ -1294,15 +1167,9 @@ async function click(selector, options = {}, ...args) {
  */
 module.exports.doubleClick = async (selector, options = {}, ...args) => {
   validate();
-  if (options instanceof RelativeSearchElement) {
-    args = [options].concat(args);
-    options = {};
-  }
-  options = {
-    waitForNavigation: options.waitForNavigation,
-    clickCount: 2,
-  };
-  await _click(selector, options, ...args);
+  options.clickCount = 2;
+  const { click } = require('./actions/click');
+  await click(selector, options, ...args);
   descEvent.emit('success', 'Double clicked ' + description(selector, true));
 };
 
@@ -1323,15 +1190,9 @@ module.exports.doubleClick = async (selector, options = {}, ...args) => {
  */
 module.exports.rightClick = async (selector, options = {}, ...args) => {
   validate();
-  if (options instanceof RelativeSearchElement) {
-    args = [options].concat(args);
-    options = {};
-  }
-  options = {
-    waitForNavigation: options.waitForNavigation,
-    button: 'right',
-  };
-  await _click(selector, options, ...args);
+  options.button = 'right';
+  const { click } = require('./actions/click');
+  await click(selector, options, ...args);
   descEvent.emit('success', 'Right clicked ' + description(selector, true));
 };
 
@@ -3165,17 +3026,6 @@ module.exports.getConfig = getConfig;
  * @param {string} [options.highlightOnAction = 'false' ] Option to highlight an element on action.
  */
 module.exports.setConfig = setConfig;
-
-const waitForMouseActions = async (options) => {
-  await doActionAwaitingNavigation(options, async () => {
-    options.type = 'mouseMoved';
-    await input.dispatchMouseEvent(options);
-    options.type = 'mousePressed';
-    await input.dispatchMouseEvent(options);
-    options.type = 'mouseReleased';
-    await input.dispatchMouseEvent(options);
-  });
-};
 
 const _focus = async (selector) => {
   let elems = [selector];
